Revert "Adding fast_float implementation."
This reverts commit cf950a79d51854695d333a336fc46c40d10efade.

+++ Include/jsonifier/RawJsonData.hpp
@@ -155,7 +155,7 @@ namespace jsonifier {
 			}
 		}
 
-		JSONIFIER_INLINE explicit operator uint64_t() const {
+		JSONIFIER_INLINE explicit operator size_t() const {
 			if (getType() == json_type::Number) {
 				return constructValueFromRawJsonData<uint64_t>(jsonData);
 			} else {
